feat: rename "checkout" module for "payment" module
