+++ config/database.js
@@ -1,10 +1,10 @@
 require('dotenv').config();
 const mongoose = require('mongoose');
 
-async function connectDB() {
+function connectDB() {
   const URI = process.env.AIRBNB_MONGO_URI;
   try {
-    await mongoose.connect(URI);
+    mongoose.connect(URI);
 
     console.log('Connected to DB');
   } catch (error) {
